+++ fs/procfs/proc_helpers.c
@@ -11,7 +11,8 @@
 #include <linux/prctl.h>  /* Definition of PR_* constants */
 #include <sys/prctl.h>
 
-#include "proc.h"
+#include "proc_helpers.h"
+
 
 #ifdef DEBUG
 # define LOG_DEBUG(fmt...) do {	\