+++ glibc/link/dl_iterate_phdr.c
@@ -7,7 +7,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "proc.h"
+#include "proc_helpers.h"
 
 
 static int callback(struct dl_phdr_info *info, size_t size, void *data)